requirements.txt
asgiref==3.2.7
Django==3.0.5
django-tenant-schemas==1.10.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
ordered-set==3.1.1
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dotenv==0.12.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
